Add regions module and move Regions classes
